requirement.txt
Flask==2.0.2
opencv-python==4.5.3.56
numpy==1.21.0
Pillow==8.2.0